legacy: BibRank CLI fix
